Setup test environment. `rake test` to run. Use stripped version of test_helper. Install test-unit gem because Ruby 2.0 only have minutest.
