+++ active_merchant_allpay.gemspec
@@ -19,6 +19,7 @@ Gem::Specification.new do |spec|
   spec.require_paths = ["lib"]
 
   spec.add_dependency 'activemerchant', '>= 1.9.4'
+  spec.add_development_dependency('test-unit', '~> 2.5.5')
 
   spec.add_development_dependency('rake')
   spec.add_development_dependency('mocha', '~> 0.13.0')
